requirement.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.0.9
djangorestframework==3.15.2
idna==3.10
mssql-django==1.5
pyodbc==5.1.0
pytz==2024.2
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.3
